* X-1.8        Camiel Vanderhoeven                             02-APR-2008  *      Fixed pointless comparison of U64 value against zero.

+++ es40/src/AlphaCPU_ieeefloat.cpp
@@ -29,6 +29,9 @@
  *
  * $Id$
  *
+ * X-1.8        Camiel Vanderhoeven                             02-APR-2008
+ *      Fixed pointless comparison of U64 value against zero.
+ *
  * X-1.7        Camiel Vanderhoeven                             14-MAR-2008
  *   1. More meaningful exceptions replace throwing (int) 1.
  *   2. U64 macro replaces X64 macro.
@@ -340,7 +343,7 @@ u64 CAlphaCPU::ieee_cvtif(u64 val, u32 ins, u32 dp)
 
   if(val == 0)
     return 0;   /* 0? return +0 */
-  if(val < 0)
+  if(((s64)val) < 0)
   {             /* < 0? */
     a.sign = 1; /* set sign */
     val = NEG_Q(val);
